requirements.txt
asgiref==3.4.1
Django==3.2.6
django-cors-headers==3.8.0
django-debug-toolbar==3.2.2
django-extensions==3.1.3
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
python-memcached==1.59
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
